+++ YumemiInternTask/Extension/WeatherErrorExtension.swift
@@ -1,0 +1,12 @@
+import Foundation
+
+extension WeatherError {
+    var errorMessage: String {
+        switch self {
+        case .invalidParameterError:
+            return "リクエストが不正です"
+        case .unknownError:
+            return "予期せぬエラーが発生しました"
+        }
+    }
+}
